correc errros CFLAGS

+++ srcs/parsing/s_quote.c
@@ -24,14 +24,14 @@ int	quote_error(char c)
 
 int	s_quote(t_copy *copy)
 {
-	if (copy->i == (ft_strlen(copy->wc) - 1))
+	if (copy->i == ((int)ft_strlen(copy->wc) - 1))
 		return (quote_error('\''));
 	if (copy->wc[copy->i + 1] == '\'' && (copy->wc[copy->i + 2] == ' '
 			|| copy->wc[copy->i + 2] == '\0'))
 		copy->cmd[++copy->j] = '\0';
 	while (copy->wc[copy->i] && copy->wc[++copy->i] != '\'')
 		copy->cmd[++copy->j] = copy->wc[copy->i];
-	if ((copy->i == ft_strlen(copy->wc)) && copy->wc[copy->i] != '\'')
+	if ((copy->i == (int)ft_strlen(copy->wc)) && copy->wc[copy->i] != '\'')
 		return (quote_error('\''));
 	copy->i++;
 	return (-2);
@@ -39,14 +39,14 @@ int	s_quote(t_copy *copy)
 
 int	s_quote_arg(t_copy *copy, size_t i)
 {
-	if (copy->i == (ft_strlen(copy->wc) - 1))
+	if (copy->i == ((int)ft_strlen(copy->wc) - 1))
 		return (quote_error('\''));
 	if (copy->wc[copy->i + 1] == '\'' && (copy->wc[copy->i + 2] == ' '
 			|| copy->wc[copy->i + 2] == '\0'))
 		copy->args[i][++copy->j] = '\0';
 	while (copy->wc[copy->i] && copy->wc[++copy->i] != '\'')
 		copy->args[i][++copy->j] = copy->wc[copy->i];
-	if ((copy->i == ft_strlen(copy->wc)) && copy->wc[copy->i] != '\'')
+	if ((copy->i == (int)ft_strlen(copy->wc)) && copy->wc[copy->i] != '\'')
 		return (quote_error('\''));
 	copy->i++;
 	return (1);
@@ -54,7 +54,7 @@ int	s_quote_arg(t_copy *copy, size_t i)
 
 int	s_quote_redir(t_copy *copy, char *str)
 {
-	if (copy->i == (ft_strlen(copy->wc) - 1))
+	if (copy->i == ((int)ft_strlen(copy->wc) - 1))
 		return (quote_error('\''));
 	if ((copy->wc[copy->i + 1] == '\'' && copy->wc[copy->i + 2] == ' ')
 		&& !str)
@@ -66,7 +66,7 @@ int	s_quote_redir(t_copy *copy, char *str)
 	}
 	while (copy->wc[copy->i] && copy->wc[++copy->i] != '\'')
 		str[++copy->redir.i] = copy->wc[copy->i];
-	if ((copy->i == ft_strlen(copy->wc)) && copy->wc[copy->i] != '\'')
+	if ((copy->i == (int)ft_strlen(copy->wc)) && copy->wc[copy->i] != '\'')
 		return (quote_error('\''));
 	str[copy->redir.i + 1] = 0;
 	copy->i++;
